Add files via upload
This code was written and used during the course of my PhD. It takes a given directory of VCF files and processes each one for unique mutation calls. It can be customized to detect any mutation at any frequency, such as shared mutations or whatnot but for my research, we were interested in only the unique mutations across cell lines. It can be optimized a bit as it does not scale well to extremely large datasets or VCF files. it can currently handle about 50 VCF files with around 1000 lines each in a few minutes currently. The scaling seems to be more a Pandas library issue, I will continue to modify this when needed.