Add files via upload
This is some code that was mostly generated by Jade Takakuwa, who at the time was a masters student helping in our lab. I helped with some logic checks and minor modifications overall but this code is part of a much larger project that was the thesis of my collegue David Patton. This code is used in conjunction with his CDMAP and CDVIZ program to calculate the early/late replication positions of a given bacteria as well as generate counts of transition/transversions and adjust for generational time of each sample.